requirements.txt
django
djangorestframework
django-environ
django-seed
psycopg2-binary